Build scripts moved to BuildScripts python project
